Fix/#120 일정 수정/삭제 권한 확인 (#121)
* [fix - #120] 데이트 일정 수정/삭제에 대한 권한 체크 추가

* [fix - #120] 개인 일정 수정/삭제에 대한 권한 체크 추가

* [fix - #120] @RequestParam 속성 수정

* [test - #120] 데이트 일정 수정/삭제 시 데이트 일정 수정 권한 체크 controller test 추가

* [test - #120] 데이트 일정 수정/삭제 시 데이트 일정 수정 권한 체크 service test 추가

* [test - #120] 개인 일정 수정/삭제 시 개인 일정 수정 권한 체크 controller test 추가

* [test - #120] 개인 일정 수정/삭제 시 개인 일정 수정 권한 체크 service test 추가

* [test - #120] 메서드 명 수정

+++ src/main/java/com/dateplan/dateplan/domain/dating/service/DatingService.java
@@ -48,6 +48,11 @@ public class DatingService {
 		}
 
 		Dating dating = datingReadService.findByDatingId(datingId);
+
+		if (isNotDatingOwner(couple.getId(), dating.getCouple().getId())) {
+			throw new NoPermissionException(Resource.DATING, Operation.UPDATE);
+		}
+
 		dating.updateDating(
 			request.getTitle(),
 			request.getLocation(),
@@ -65,10 +70,19 @@ public class DatingService {
 		}
 
 		Dating dating = datingReadService.findByDatingId(datingId);
+
+		if (isNotDatingOwner(couple.getId(), dating.getCouple().getId())) {
+			throw new NoPermissionException(Resource.DATING, Operation.DELETE);
+		}
+
 		datingRepository.delete(dating);
 	}
 
 	private boolean isNotSameCouple(Long requestId, Long coupleId) {
 		return !Objects.equals(requestId, coupleId);
+	}
+
+	private boolean isNotDatingOwner(Long coupleId, Long datingOwnerId) {
+		return !Objects.equals(coupleId, datingOwnerId);
 	}
 }
